backend/monolithic_app/requirements.txt
# Django and core dependencies
Django>=5.2.4

# Database
psycopg2-binary>=2.9.10

# Environment management
django-environ>=0.12.0

# Django REST Framework
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
django-filter>=25.1

# Authentication and Authorization
dj-rest-auth>=7.0.1
django-allauth>=65.9.0

# Security
cryptography>=45.0.5

# Background Tasks
celery[redis]>=5.5.3
redis>=6.2.0

# Static files and media
whitenoise>=6.9.0
pillow>=11.3.0

# Utilities
django-extensions>=4.1
requests>=2.32.4

# CORS headers for frontend integration
django-cors-headers>=4.0.0

# Production server
gunicorn>=22.0.0

# Development dependencies (optional)
django-debug-toolbar>=4.0.0
